refactor: переместить state в пакет service
