+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -6,8 +6,8 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.booking.dto.BookingCreationDto;
 import ru.practicum.shareit.booking.dto.BookingFullDto;
-import ru.practicum.shareit.booking.model.State;
 import ru.practicum.shareit.booking.service.BookingService;
+import ru.practicum.shareit.booking.service.State;
 import ru.practicum.shareit.validation.util.ValidationUtil;
 
 import javax.validation.Valid;
