+++ src/main/java/ru/practicum/shareit/booking/service/State.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking.model;
+package ru.practicum.shareit.booking.service;
 
 public enum State {
     ALL,
